DGEN2.jar reinstantiates GT probability class for case found of bug when used for many taxa
Found internally that gene tree probabilities were not being calculated correctly in a case with many taxa. This was due to not reinstantiating the class that calculates the gene tree probabilities before passing in a network with new branch lengths. The class requires reinstantiating because of its typical use case where it saves things in an internal state.